Add support for ScrivePdfTools Service
Kontrakcja relies on ScrivePdfTools for processing PDF files. When
called as a process, the subsequent process of starting the JVM incurs
a delay of more than half a second for every call to ScrivePdfTools.

ScrivePdfTools already contains almost everything needed. This patch
just adds remove-javascript to the list of admissible commands and
updates some outdated code.